authenticating tasks routes

+++ src/routers/task.js
@@ -23,32 +23,32 @@ router.post('/tasks', auth, async (req, res) => {
 });
 
 //Fetching all tasks
-router.get('/tasks', async (req, res) => {
-    const task = await Task.find({});
+router.get('/tasks', auth, async (req, res) => {
+    // const task = await Task.find({});
     try {
-        res.status(200).send(task)
+        await req.user.populate('tasks').execPopulate()
+        res.status(200).send(req.user.tasks)
     } catch (e) {
         res.status(500).send(e)
     }
 });
 
 //Fetch single task
-router.get('/tasks/:id', async (req, res) => {
+router.get('/tasks/:id', auth, async (req, res) => {
     const _id = req.params.id;
     try {
-        const tasks = await Task.findById(_id);
-        if (!tasks) {
-            return res.status(400).send()
-        } else {
-            res.status(200).send(tasks)
+        const task = await Task.findOne({ _id, owner: req.user._id })
+        if (!task) {
+            return res.status(404).send()
         }
+        res.send(task)
     } catch (e) {
         res.status(500).send(e)
     }
 });
 
 //Updating a task
-router.patch('/tasks/:id', async (req, res) => {
+router.patch('/tasks/:id', auth, async (req, res) => {
     const updates = Object.keys(req.body)
     const allowedUpdates = ['description', 'completed'];
     const isValidOperation = updates.every((update) => {
@@ -60,19 +60,14 @@ router.patch('/tasks/:id', async (req, res) => {
     }
 
     try {
-        const task = await Task.findById(req.params.id);
+        const task = await Task.findOne({ _id: req.params.id, owner: req.user._id })
+        if (!task) {
+            return res.status(404).send()
+        }
         updates.forEach((update) => {
             task[update] = req.body[update];
         })
         await task.save();
-
-        // const task = await Task.findByIdAndUpdate(req.params.id, req.body, {
-        //     new: true,
-        //     runValidators: true
-        // })
-        if (!task) {
-            return res.status(400).send()
-        }
         res.send(task)
     } catch (e) {
         res.status(400).send(e)
@@ -80,11 +75,11 @@ router.patch('/tasks/:id', async (req, res) => {
 });
 
 //Deleting task
-router.delete('/tasks/:id', async (req, res) => {
+router.delete('/tasks/:id', auth, async (req, res) => {
     try {
-        const task = await Task.findByIdAndDelete(req.params.id);
+        const task = await Task.findByIdAndDelete({ _id: req.params.id, owner: req.user._id });
         if (!task) {
-            return res.status(400).send();
+            return res.status(404).send();
         }
         res.send(task)
     } catch (e) {
